feat: move languageDefault to monaco.contribution
